initial-cleanup-authsession-methods
