+++ app/src/main/java/org/ole/planet/takeout/DashboardFragment.java
@@ -1,6 +1,5 @@
 package org.ole.planet.takeout;
 
-import android.app.IntentService;
 import android.app.ProgressDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -15,7 +14,6 @@ import android.support.v4.app.Fragment;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.app.AlertDialog;
 import android.text.TextUtils;
-import android.util.Base64;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -25,54 +23,25 @@ import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import com.google.android.flexbox.FlexDirection;
 import com.google.android.flexbox.FlexboxLayout;
-
-
-import org.json.JSONObject;
 import org.ole.planet.takeout.Data.Download;
 import org.ole.planet.takeout.Data.realm_myCourses;
-
-import org.ole.planet.takeout.ExoPlayerVideo;
 import org.ole.planet.takeout.Data.realm_offlineActivities;
-import org.ole.planet.takeout.callback.OnHomeItemClickListener;
-import org.ole.planet.takeout.datamanager.ApiClient;
-import org.ole.planet.takeout.datamanager.ApiInterface;
-import org.ole.planet.takeout.datamanager.MyDownloadService;
-import org.ole.planet.takeout.library.MyLibraryFragment;
-import org.ole.planet.takeout.userprofile.UserProfileDbHandler;
-import org.ole.planet.takeout.userprofile.UserProfileFragment;
 import org.ole.planet.takeout.utilities.DialogUtils;
 import org.ole.planet.takeout.utilities.Utilities;
-
-import java.io.DataOutputStream;
 import java.io.File;
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Timer;
-import java.util.TimerTask;
-
 import org.ole.planet.takeout.Data.realm_myLibrary;
-import org.ole.planet.takeout.datamanager.MyDownloadService;
-
 import java.util.UUID;
-
-
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
 import io.realm.RealmResults;
-import okhttp3.ResponseBody;
-import retrofit2.Call;
-import retrofit2.Response;
-
 import static android.content.Context.MODE_PRIVATE;
 import static org.ole.planet.takeout.Dashboard.MESSAGE_PROGRESS;
+import org.ole.planet.takeout.AuthSessionUpdater;
 
 /**
  * A placeholder fragment containing a simple view.
@@ -81,7 +50,7 @@ public class DashboardFragment extends Fragment {
 
     //TextViews
     public static final String PREFS_NAME = "OLE_PLANET";
-    SharedPreferences settings;
+    public static SharedPreferences settings;
     TextView txtFullName, txtCurDate, txtVisits;
     String fullName;
     Realm mRealm;
@@ -94,10 +63,7 @@ public class DashboardFragment extends Fragment {
     private ImageButton myTeamsImage;
     public String globalFilePath = Environment.getExternalStorageDirectory() + File.separator + "ole" + File.separator;
 
-    private Call<ResponseBody> request;
-
-    private String auth = ""; // Main Auth Session Token for any Online File Streaming/ Viewing -- Constantly Updating Every 15 mins
-
+    private static String auth = ""; // Main Auth Session Token for any Online File Streaming/ Viewing -- Constantly Updating Every 15 mins
 
     private BroadcastReceiver broadcastReceiver = new BroadcastReceiver() {
         @Override
@@ -143,7 +109,7 @@ public class DashboardFragment extends Fragment {
         myLibraryDiv(view);
         myCoursesDiv(view);
         showDownloadDialog();
-        timerSendPostNewAuthSessionID();
+        AuthSessionUpdater.timerSendPostNewAuthSessionID(settings);
     }
 
     public int offlineVisits() {
@@ -216,6 +182,20 @@ public class DashboardFragment extends Fragment {
         }
     }
 
+    public void setTextViewProperties(TextView[] textViewArray, int itemCnt, realm_myLibrary items, realm_myCourses itemsCourse) {
+        textViewArray[itemCnt] = new TextView(getContext());
+        textViewArray[itemCnt].setPadding(20, 10, 20, 10);
+        textViewArray[itemCnt].setBackgroundResource(R.drawable.dark_rect);
+        textViewArray[itemCnt].setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
+        textViewArray[itemCnt].setGravity(Gravity.CENTER_VERTICAL | Gravity.CENTER_HORIZONTAL);
+        textViewArray[itemCnt].setTextColor(getResources().getColor(R.color.dialog_sync_labels));
+        if(items != null){
+            textViewArray[itemCnt].setText(items.getTitle());
+        }else if(itemsCourse != null){
+            textViewArray[itemCnt].setText(itemsCourse.getCourseTitle());
+        }
+    }
+
     private void showDownloadDialog() {
         final RealmResults<realm_myLibrary> db_myLibrary = mRealm.where(realm_myLibrary.class).equalTo("resourceOffline", false).findAll();
         if (!db_myLibrary.isEmpty()) {
@@ -251,7 +231,6 @@ public class DashboardFragment extends Fragment {
         ArrayList urls = new ArrayList();
         for (int i = 0; i < selectedItems.size(); i++) {
             urls.add(Utilities.getUrl(db_myLibrary.get(selectedItems.get(i)), settings));
-            Log.e("URLS",""+urls);
         }
         startDownload(urls);
     }
@@ -260,6 +239,16 @@ public class DashboardFragment extends Fragment {
         if (!urls.isEmpty()) {
             prgDialog.show();
             Utilities.openDownloadService(getActivity(), urls);
+        }
+    }
+
+    public void setProgress(Download download) {
+        prgDialog.setProgress(download.getProgress());
+        if (!TextUtils.isEmpty(download.getFileName())) {
+            prgDialog.setTitle(download.getFileName());
+        }
+        if (download.isCompleteAll()) {
+            DialogUtils.showError(prgDialog, "All files downloaded successfully");
         }
     }
 
@@ -285,31 +274,7 @@ public class DashboardFragment extends Fragment {
         });
     }
 
-    public void setTextViewProperties(TextView[] textViewArray, int itemCnt, realm_myLibrary items, realm_myCourses itemsCourse) {
-        textViewArray[itemCnt] = new TextView(getContext());
-        textViewArray[itemCnt].setPadding(20, 10, 20, 10);
-        textViewArray[itemCnt].setBackgroundResource(R.drawable.dark_rect);
-        textViewArray[itemCnt].setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
-        textViewArray[itemCnt].setGravity(Gravity.CENTER_VERTICAL | Gravity.CENTER_HORIZONTAL);
-        textViewArray[itemCnt].setTextColor(getResources().getColor(R.color.dialog_sync_labels));
-        if(items != null){
-            textViewArray[itemCnt].setText(items.getTitle());
-        }else if(itemsCourse != null){
-            textViewArray[itemCnt].setText(itemsCourse.getCourseTitle());
-        }
-    }
-
-    public void setProgress(Download download) {
-        prgDialog.setProgress(download.getProgress());
-        if (!TextUtils.isEmpty(download.getFileName())) {
-            prgDialog.setTitle(download.getFileName());
-        }
-        if (download.isCompleteAll()) {
-            DialogUtils.showError(prgDialog, "All files downloaded successfully");
-        }
-    }
-
-    private void openFileType(final realm_myLibrary items, String videotype){
+    public void openFileType(final realm_myLibrary items, String videotype){
         if(items.getMediaType().equals("video")) {
             playVideo(videotype, items);
         }else{
@@ -317,109 +282,49 @@ public class DashboardFragment extends Fragment {
         }
     }
 
+    //Sets Auth Session Variable every 15 mins
+    public void setAuthSession(Map<String, List<String>> responseHeader){
+        String headerauth[] = responseHeader.get("Set-Cookie").get(0).split(";");
+        auth = headerauth[0];
+    }
+
     // Plays Video Using ExoPlayerVideo.java
-    private void playVideo(String videoType, final realm_myLibrary items){
+    public void playVideo(String videoType, final realm_myLibrary items){
         Intent intent = new Intent(DashboardFragment.this.getActivity(), ExoPlayerVideo.class);
         Bundle bundle = new Bundle();
         bundle.putString("videoType", videoType);
         if(videoType.equals("online")){
             bundle.putString("videoURL",""+items.getResourceRemoteAddress());
-            bundle.putString("Auth", auth);
+            Log.e("AUTH",""+auth);
+            bundle.putString("Auth", ""+auth);
         }else if(videoType.equals("offline")){
-            bundle.putString("videoURL",""+ Uri.fromFile(new File(""+Utilities.getSDPathFromUrl(items.getResourceRemoteAddress()))));
+            bundle.putString("videoURL",""+ Uri.fromFile(new File(""+ Utilities.getSDPathFromUrl(items.getResourceRemoteAddress()))));
             bundle.putString("Auth", "");
         }
         intent.putExtras(bundle);
         startActivity(intent);
     }
 
-    // sendPost() - Meant to get New AuthSession Token for viewing Online resources such as Video, and basically any file.
-    // It creates a session of about 20mins after which a new AuthSession Token will be needed.
-    // During these 20mins items.getResourceRemoteAddress() will work in obtaining the files necessary.
-    public void sendPost() {
-        Thread thread = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    HttpURLConnection conn = (HttpURLConnection) getSessionUrl().openConnection();
-                    conn.setRequestMethod("POST");
-                    conn.setRequestProperty("Content-Type", "application/json");
-                    conn.setRequestProperty("Accept","application/json");
-                    conn.setDoOutput(true);
-                    conn.setDoInput(true);
-
-                    DataOutputStream os = new DataOutputStream(conn.getOutputStream());
-                    os.writeBytes(getJsonObject().toString());
-
-                    os.flush();
-                    os.close();
-
-                    setAuthSession(conn.getHeaderFields());
-                    conn.disconnect();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        thread.start();
+    public void openIntent(realm_myLibrary items, Class typeClass) {
+        Intent fileOpenIntent = new Intent(DashboardFragment.this.getActivity(), typeClass);
+        fileOpenIntent.putExtra("TOUCHED_FILE", items.getResourceLocalAddress());
+        startActivity(fileOpenIntent);
     }
 
-    public void setAuthSession(Map<String, List<String>> responseHeader){
-        String headerauth[] = responseHeader.get("Set-Cookie").get(0).split(";");
-        auth = headerauth[0];
-    }
-
-    // Updates Auth Session Token every 15 mins to prevent Timing Out
-    public void timerSendPostNewAuthSessionID() {
-        Timer timer = new Timer ();
-        TimerTask hourlyTask = new TimerTask () {
-            @Override
-            public void run () {
-                sendPost();
-            }
-        };
-        timer.schedule(hourlyTask, 0, 1000*60*15);
-    }
-
-    public JSONObject getJsonObject(){
-        try {
-            JSONObject jsonParam = new JSONObject();
-            jsonParam.put("name", settings.getString("url_user",""));
-            jsonParam.put("password", settings.getString("url_pwd",""));
-            return jsonParam;
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return null;
-    }
-    public URL getSessionUrl(){
-        try {
-            String pref = settings.getString("serverURL", "");
-            pref += "/_session";
-            URL SERVER_URL = new URL(pref);
-            return SERVER_URL;
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return null;
-    }
     public void checkFileExtension(realm_myLibrary items) {
         String filenameArray[] = items.getResourceLocalAddress().split("\\.");
         String extension = filenameArray[filenameArray.length - 1];
 
         switch (extension) {
             case "pdf":
-                Intent pdfOpenIntent = new Intent(DashboardFragment.this.getActivity(), PDFReaderActivity.class);
-                openIntent(items, pdfOpenIntent);
+                openIntent(items,PDFReaderActivity.class);
                 break;
             case "bmp":
             case "gif":
             case "jpg":
             case "png":
             case "webp":
-                Intent imageOpenIntent = new Intent(DashboardFragment.this.getActivity(), ImageViewerActivity.class);
-                openIntent(items, imageOpenIntent);
+                openIntent(items,ImageViewerActivity.class);
                 break;
             default:
                 checkMoreFileExtensions(extension, items);
@@ -427,28 +332,20 @@ public class DashboardFragment extends Fragment {
         }
     }
 
-    private void openIntent(realm_myLibrary items, Intent fileOpenIntent) {
-        fileOpenIntent.putExtra("TOUCHED_FILE", items.getResourceLocalAddress());
-        startActivity(fileOpenIntent);
-    }
-
     public void checkMoreFileExtensions(String extension, realm_myLibrary items)
     {
         switch (extension) {
             case "txt":
-                Intent textFileOpenIntent = new Intent(DashboardFragment.this.getActivity(), TextFileViewerActivity.class);
-                openIntent(items, textFileOpenIntent);
+                openIntent(items,TextFileViewerActivity.class);
                 break;
             case "md":
-                Intent markdownOpenIntent = new Intent(DashboardFragment.this.getActivity(), MarkdownViewerActivity.class);
-                openIntent(items, markdownOpenIntent);
+                openIntent(items,MarkdownViewerActivity.class);
                 break;
             case "csv":
-                Intent CSVOpenIntent = new Intent(DashboardFragment.this.getActivity(), CSVViewerActivity.class);
-                openIntent(items, CSVOpenIntent);
+                openIntent(items,CSVViewerActivity.class);
                 break;
             default:
-                Toast.makeText(getContext(), "This file type is currently unsupported", Toast.LENGTH_LONG).show();
+                Toast.makeText(DashboardFragment.this.getContext(), "This file type is currently unsupported", Toast.LENGTH_LONG).show();
                 break;
         }
